chore(smart-wallet): export contractExports test tools
